Added StrategyController
