+++ Presenentation/Hypance.WebApi/Controllers/SymbolController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Hypance.Core.Domain.Strategies;
 using Hypance.Core.Domain.Symbols;
 using Hypance.Data;
 using Microsoft.AspNetCore.Mvc;
@@ -23,36 +24,12 @@ namespace Hypance.WebApi.Controllers
 
         // GET: api/values
         [HttpGet]
-        public IEnumerable<Symbol> Get()
+        public IEnumerable<Symbol> GetAll()
         {
             return _symbolRepository.GetAll();
         }
 
-        // GET api/values/5
-        [HttpGet("{id}")]
-        public string Get(int id)
-        {
-            return "value";
-        }
-
-        // POST api/values
-        [HttpPost]
-        public void Post([FromBody]string value)
-        {
-
-        }
-
-        // PUT api/values/5
-        [HttpPut("{id}")]
-        public void Put(int id, [FromBody]string value)
-        {
-        }
-
-        // DELETE api/values/5
-        [HttpDelete("{id}")]
-        public void Delete(int id)
-        {
-        }
+        
     }
 }
